fix(pagination): the current page is a page number, so make it one
CURRENT_PAGE was a float, and an off-boundary offset made it fractional:
?start=30 with 20 per page exported 2.5, which
`{if $page == $navbar->currentPage}` matched against no key in the pages
list at all. The bar rendered with nothing highlighted.

The history says this was collateral, not intent. The 2009 original
computed `ceil($start / $per) + 1` -- a whole number, assigned straight to
CURRENT_PAGE. 856b5a2519 (2012) replaced it while doing something else
entirely: its subject was "navigation bar urls are always on the number of
images per page boundaries", a Googlebot crawl fix. It added the round()
snap for the URLs, needed a fractional position for its new
floor()/ceil() window bounds, and reused one variable for both jobs.

The definition now is the page the FIRST displayed element falls on --
`min(intdiv($start, $nbElementPage) + 1, $maximum)` -- and everything else
derives from it: the links are that page's neighbours, and the window is
centred on it, which is what $pages_around has always claimed to mean.
Navbar::$currentPage is ?int and both navigation_bar templates read
`$page === $navbar->currentPage` with no cast.

Two defects fall out that the fraction was hiding:

- A "next" link pointing at the page already shown. At start=85 of 100 the
  unrounded 5.25 !== 5, so the next/last block ran on the last page, and
  URL_NEXT then had to be capped back to $last.
- An out-of-range offset naming a page that does not exist. $start is
  never checked against the element count, so ?start=500 on a 5-page
  gallery reported page 26 -- absent from the list, leaving the template
  exactly where the fraction left it. Hence the clamp.

`$next < $last ? $next : $last` goes with them: inside
`$currentPage !== $maximum` and with the clamp above, $next cannot exceed
$last. It only ever guarded the out-of-range case now stopped at source.

Validation, adversarial:

- Old vs new over 15,306 inputs (4 page sizes x 5 element counts x 3
  pages_around x every offset to nb + 2*per), comparing all 7 fields:
  ZERO differences on any boundary-aligned offset. Every URL Piwigo emits
  is boundary-aligned, so normal navigation is byte-identical. Off
  boundary it is a different navbar, not just a different label --
  URL_PREV differs in 5181 inputs, pages in 7324, URL_NEXT in 4833.
- The independent argument for that: round() made "next" SKIP content.
  Viewing elements 30-49, next jumped to 60, leaving 50-59 unreachable by
  clicking. floor overlaps but never skips.
- 8 invariants over 23,535 cases, 0 failures. Degenerate inputs behave:
  PHP_INT_MAX start clamps to the last page, per=0 throws
  DivisionByZeroError in old and new alike, and intdiv's one hazard
  (PHP_INT_MIN / -1) is unreachable behind the existing `$start < 0` clamp.

The test file's mutation notes are corrected rather than left to rot. Two
"confirmed-equivalent" claims are no longer true and now say so: the
`(int)` casts and `$previous > 0` are killed, the casts because intdiv()
declares int parameters under strict_types where float division coerced a
numeric string. The notes describing deleted code are gone, and every
`line NN's` reference is replaced by the expression it names, since the
numbers had all shifted and would shift again.

`pest --mutate` scores this class 100%, and that is not to be read as
proof: `$start <= 0` is reported KILLED but survives the entire Unit
suite, hand-verified. It is a genuinely equivalent mutant, so no gap --
but the score overstates.

Also repairs a red test this campaign shipped: FilterPanelRendererTest
asserted `getTemplateVars('filter')` is `[]`, which stopped being true
when 29ab56c82e handed the template BulkManagerFilter instead of the raw
session array. That commit was verified with the BatchManager Browser
suites, golden-html and VR, but not the Unit suite, where the renderer's
own test lives.

Unit/Arch 5575, PHPStan (both gates), ECS, golden-html 91 and VR 82 with
nothing regenerated, Comments/BatchManagerUnit/Rating Browser 60.

+++ src/Piwigo/Core/Projection/Navbar.php
@@ -23,7 +23,7 @@ final readonly class Navbar
      * @param array<int, string> $pages
      */
     public function __construct(
-        public ?float $currentPage = null,
+        public ?int $currentPage = null,
         public ?string $urlFirst = null,
         public ?string $urlPrev = null,
         public ?string $urlNext = null,
@@ -43,7 +43,7 @@ final readonly class Navbar
      * mutation-tested coverage with it) -- this is the one conversion
      * point at the very end, from that raw legacy shape to this VO.
      *
-     * @param array{CURRENT_PAGE?: float, URL_FIRST?: string, URL_PREV?: string, URL_NEXT?: string, URL_LAST?: string, pages?: array<int, string>, NB_PAGE?: int} $row
+     * @param array{CURRENT_PAGE?: int, URL_FIRST?: string, URL_PREV?: string, URL_NEXT?: string, URL_LAST?: string, pages?: array<int, string>, NB_PAGE?: int} $row
      */
     public static function fromLegacyArray(array $row): self
     {
